Bug 371586 - fixed sort by creation time and added more diagnostics
MarkerCreationTimeField.compare() was broken because it could return
integer values breaking the transitive comparison of long values.

The comparator used int values downcasted from longs computed as
difference between long timestamps. If the difference values were
outside of the Integer range, the downcast could produce integers with a
different sign than the original long value. So for the long values
A>B>C comparator could return (A-B) > 0, (B-C) > 0 but (A-C) < 0 etc,
which was of course wrong and caused errors in TimSort.

Beside this fix, added extra diagnostics for the other possible unstable
sorter cases in Markers.sortMarkerEntries, and added (silent) error log
handler to avoid not really helpful error dialogs.

Change-Id: I5ee4876615a448e502ab2b61df06d91230d92991
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>