Bug 531486 - Progress dialog does not update progress of all jobs
Whenever the layout of a DetailedProgressViewer is computed, the
visibility of the progress items in the viewer is updated. Computing the
layout of an item itself however, does not trigger a visibility update.
This results in cases where a progress item is not yet spanned, but its
parent DetailedProgressViewer already sets its visibility (based on its
location and bounds). Since the item bounds are not set yet, as its not
spanned, the item is perceived as not visible.

We therefore observe visible progress items which don't update their
progress, i.e. some jobs appear as stuck. As soon as the viewer is
resized or scrolled, the visibility of the items is correctly set (since
they are spanned at this point) and progress is shown correctly.

This change ensures that the visibility of progress items is also
updated when they, themselves, are resized (e.g. due to a layout
request). This results in correct progress shown for jobs displayed in a
DetailedProgressViewer.

Change-Id: I3819c54221ca85239254c39da7b416e72b88f44c
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>