Bug 530330 - SafeChunkyOutputStream.close() does not flush() on Java9
SafeChunkyOutputStream extends FilterOutputStream and in Java 9, the
close() changed that it does not allow to call it twice. However
SafeChunkyOutputStream code is funky enough to allow re-use of the
stream (it opens new underlined stream).

The fix is to override close() and go back to the Java 8 implementation
of close().

Change-Id: I85954037b0eda47005f883fa4f461c5f79a27deb
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>