Bug 532579 - Center prototype tab icon
The prototype tab icon is aligned at the bottom of the 16x16 area
instead of in the center. This leads to it not being aligned with the
remaining tab icons in the launch config UI.

In the attached images the icon has been centered again. I'm aware a
separate change is needed for eclipse.platform.images.

Change-Id: I477589d3073f91f25c9a39f906fb5e841514478c
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>