Bug 532706 - Update minings when code mining manager is initialized
After fCodeMiningManager is initialized in
ensureCodeMiningManagerInstalled, it is not assured that
updateCodeMinings() is called afterwards. The initial reconciling has
already called before an annotation painter is set. But only when this
has happened the code mining is properly installed and an initial mining
has to be shown. Thus ensureCodeMiningManagerInstalled calls
updateCodeMinings after initialization of the manager.

Also changed guard conditions since it is intended to check that the
code mining manager is set, but method hasCodeMiningProviders is from
its name not so clear that this is implied by calling it.

Change-Id: Ic79743178aa3302d38586770246ca7d749401e15
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>