Bug 533927 - Move o.e.ui.cocoa to Java 8
In an effort to move to Tycho 1.2.0 I've uncovered this one caused by
"As Tycho use bundle BREE to provision a EE during dependency
resolution, every bundle that depends on other bundles with higher BREE
will see a similar error, because newer EE will be missing to transitive
deps during resolution. Either such "old BREE" bundles have to bump
their BREE to keep in sync, or the pom.xml needs to be tweaked to force
a newer EE at build-time for these bundles (or for the whole project),
or target-platform-configuration must be tweaked to ignore EE
requirements. " as can be seen in the release notes
(https://wiki.eclipse.org/Tycho/Release_Notes/1.2) . Discussions on
Tycho side howto prevent this are ongoing but this should not block our
adoption as this change at least improves consistency.

Change-Id: Ie381bf473faa36d03cb0e7e1da2a23921d518def
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>