Bug 534803 - Migration org.eclipse.e4.ui.tests.css.core to JUnit5
Version bump for 4.10

Added dependencies on JUnit 5 libraries on compile-time dependencies
only. Runtime dependency on jupiter engine is added by JUnit 5 classpath
container. A compile time dependency can lead to problems with
tycho-surefire plugin (see bug#534587).

Refactored test classes to use JUnit 5 API.

StyledRuleTest: assertEquals for floats with delta of 0f is not valid in
JUnit5 Assertions. When a delta is specified, it must be a positive
value.

Change-Id: Ia6c6db1ac2f84cd80c0ba16402071f25f68954b6
Also-by: Karsten Thoms <karsten.thoms@itemis.de>
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>