Bug 535454 - do not allow null context in DebugContextEvent constructor
In case of a null context passed to the constructor of
DebugContextEvent, a NPE could be thrown later on when the event is
being processed. This is the case since DebugContextEvent.getContext()
makes no mention of possibly returning a null. E.g. some code in
ToggleStepFiltersAction and LaunchView does not check if the returned
context is null.

At the point of throwing the NPE, its unclear where the null context
actually came from. For issues which are difficult to reproduce, this
information is vital.

With this change, an exception is thrown if the passed context is null.

Change-Id: If7a86500bc827bc6aa73ca971c9be79a714be81f
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>