Bug 532530 - Eclipse Forms no longer wrapping
Add a new test suite, ScrolledFormTest, that validates the
size of the content area after laying out ScrolledForms
with various parameters.

Change SharedScrollComposite.reflow to:
- Apply the minimum size in a manner closer to what
  4.6 used.
- Handle the expandHorizontal=false case in a manner closer
  to what the JavaDoc promises (that is: it won't stretch a
  control beyond its preferred size in cases where the
  expand flag is disabled).

Update SizeCache.computeMinimumWidth to handle SWT wrapping
controls better. Uses the same heuristics as
LayoutGenerator for identifying wrapping controls. For
wrapping controls with very long preferred sizes, the
new heuristics clamp the horizontal preferred size.
This is safe to do since such controls can always wrap
to avoid truncation and the purpose of the
computeMinimumWidth method is to compute the minimum
usable size for the control without truncation.

There are a lot of corner cases, so it may be easier
to read the test cases to understand what the
intended behavior is than to try to figure out the
behavior by looking at the code itself.

Change-Id: I41e508710272fe36f0208c8e6eff5de6bff94de9