Bug 535462 - NPE when closing second window during debug
The bug occurs while debugging, having two windows and a custom debug
context provider. When closing the second window,
DebugWindowContextService.notify is called with the selection of the
custom debug context provider. If this selection is e.g. a stack frame,
SourceLookupService asks SourceLookupFacility to open a source file.
This throws a NPE, since the workbench page window is closing.

With this change SourceLookupFacility checks if the active workbench
window is closing. If so, the source is not opened, avoiding the NPE.

Note: updated platform.ui version range to see re-export updated
ui.workbench API.

Change-Id: Ifb177c1211248b0cda17c6d14201a210d5fec8dc
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>