Bug 531854 - View toolbar misses commands with visibleWhen expression
The problem is seen when adding a command to a view toolbar, and using a
visibleWhen with variable expression. If the command is not visible by
default, but becomes visible later on, the command will not become
visible in the toolbar of the view.

This results from using RowData.exclude on the view menu toolbar. The
view menu toolbar is not visible if there are no view menu items. This
is implemented by using RowData.exclude. When a visibleWhen expression
evaluates true, the composite of the view menu and command toolbars is
not updated correctly.

This change adds a missing pack() call to the parent composite, fixing
the problem.

Change-Id: Ia0aa31745c33ad5c8050bab484a551589b28bf11
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>