Bug 535679: load quick access elements in a job
The quick access field (top right corner of Eclipse UI) loads previously
called elements on creation. This has two drawbacks. First, for
previously called commands, handlers of all defined commands are loaded.
And second, loading an element can e.g. involve IO and so block the UI
for a substantial amount of time.

This change moves loading of elements to a job. The loading is indicated
by a message in the quick access dialog. The job is provided with a
progress monitor that can be cancelled. If cancelled, not all quick
access elements will be restored.

Furthermore CommandProvider is adjusted to not load all commands when a
specific command is requested via
CommandProvider.getElementForId(String).

Change-Id: Ie3ef2382178b3d8f32493b120f750c004561f8e1
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>