Bug 531854 - View toolbar misses commands with visibleWhen expression
The problem is seen when adding a command to a view toolbar, and using a
visibleWhen with variable expression. If the command is not visible by
default, but becomes visible later on, the command will not become
visible in the toolbar of the view.

This results from using RowData.exclude on the view menu toolbar. The
view menu toolbar is not visible if there are no view menu items. This
is implemented by using RowData.exclude. When a visibleWhen expression
evaluates true, the composite of the view menu and command toolbars is
not updated correctly.

This change adds a missing pack() call to the parent composite, fixing
the problem.

Note, that to avoid issues with broken complex layouts after the toolbar
is updated (see bug 536409), we change layout() to requestLayout(),
which runs *after* all necessary children updates are performed due
pack().

Change-Id: Ib2225ef8dfbb6a106e4853f8e8b01f716d4620a1
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Also-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>