Bug 536648 - Thread evaluation runs in parallel with thread label update
When running an evaluation on a thread suspension, its possible that a
label update of the Debug View runs in parallel. The label update can
then retrieve the intermediate state of the evaluation. If no further
label update is issued, the thread will be shown as running until the
Debug View is refreshed.

With this change, suspending a thread after an evaluation will also
update  the thread state in the Debug View. This avoids the wrong
running icon of the thread.

Change-Id: Id6ec6a9d48c041d0b2617adb8b7da91ba5140b15
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>