Bug 536952 - Avoid QuickDiff re-enabling in DocumentRewriteSessionEvent
* Only resume the DocumentLineDiffer if if was previously suspended by a
DocumentRewriteSessionEvent.
* Add a null check at
AbstractDocument.stopRewriteSession(DocumentRewriteSession) to avoid a
NullPointerException if startRewriteSession is not called before calling
stopRewriteSession. The javadoc of stopRewriteSession says "This method
has only any effect if startRewriteSession has been called before.",
without the null-check the method had an effect if startRewriteSession
was not called before, namely an NPE.

Change-Id: Ia6f92c686c934e1c7cbb514e2c32ac055b557935
Signed-off-by: Etienne Reichenbach <etienner@gmail.com>