Bug 536677 - dispose WorkbenchMenuService contributions on part disposal
This is a workaround for an implementation (or design) issue of
WorkbenchMenuService, see bug 537046, which can contribute menus and
toolbars to parts, but the link from contribution manager to the created
menu/toolbar model is lost (overridden) later. Since fixing this is
probably not trivial, we are just trying to minimize the damage.

On part disposal we now ask WorkbenchMenuService to clear all
contributed items to part menu and toolbar.

Change-Id: Ifd5d202e03ce1b6690613409d1350fddcba2688a
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>