Bug 537131 - Preference to hide Quick Access field per default
This change adds a new preference in the org.eclipse.ui.workbench
plug-in. Setting this preference to true will hide the quick access
search field, if the field state was not restored explicitly by the
user.

I.e. if the preference is used, starting in a new workspace will result
in hidden quick access field. If the field is restored at least once,
the preference is no longer used.

The preference can be set with the following entry:

org.eclipse.ui/hideQuickAccessPerDefault=true

Change-Id: I9d0d628e00503bab10f9823177164d28e2eda5f7
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>