Bug 536677 - remove all contributions on disposing toolbar
ToolBarManager.dispose() disposes children but does not releases them.
This is supposed to be done by removeAll(), but almost everywhere the
code was not doing that.

Changing ToolBarManager.dispose() would be a better solution, but it
looks like all subclasses of ContributionManager except MenuManager
don't do that and explicitly mention that removeAll() should be used.

Change-Id: I8f2240463616af4b00f5bef6257154ef6c528079
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>