Bug 537322 - Improved logging in exception case of flushMarkerCache
Whenever an exception occurs during marker attribute setting in
BreakpointWorkingSetCache.flushMarkerCache, the logged error does not
include information for the attributes which are to be set on the
marker. When logged e.g. from tests, its difficult to understand where
the logged error is coming from.

This change adds some extra information to the logged error.

Change-Id: I0b9a1aadb9310b26778f200df4004cdbb8756eef
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>