Bug 537330 - Breakpoints View NPE on perspective switch during debug
The error log indicates that at the time of
BreakpointsView.contextActivated, the Breakpoints View had no tree model
viewer. This caused a NPE to be logged in the Error Log.

This change adds a guard to the method, similar to the guard found in
VariablesView.contextActivated. This avoids the logged NPE and does not
alter behavior otherwise; the rest of BreakpointsView.contextActivated
does nothing if there is no tree model.

Change-Id: I070a2714c40636fb8796a963b7a3d4c147c04bf9