Bug 536207: Fix the height calculation by using the bottom index.
To fix the height of the displayed viewport instead of trying to
calculate the height the same approach as for the top index is used.
This way the height is ok for all cases including enabled word wrap.

Also fixed the update behavior when the text changes and introduced a
minimal height of the viewport rectangle, such that it does not shrink
if content is shown completely in the text viewer.

Change-Id: Ife949b9db6a7cbb46be749d082b40bc843279a79
Signed-off-by: Arne Deutsch <Arne.Deutsch@itemis.de>