Bug 538548 - Debug View in another perspective causes editor to pop-up
If a window has the Debug View open in an inactive perspective,
breakpoint hits will cause an editor to open in that window. This
hinders the use of multiple windows, each with a different type of
perspective. Only windows which have the Debug View in their active
perspective should show source on breakpoint hit.

The editor is activated due to the Debug View reacting to selection
events, which occur due to breakpoint hits.

This change ensures that the Debug View will react to selection events
only if its active. I.e. only if the active perspective of the window
contains the Debug View.

Change-Id: Ida92f5322e1a83b95af4bc43f25a9fac9d4519b6
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>