Bug 538989 - FormText: Copying selection text contains unwanted leading newline
When the selection is computed in FormText#ComputeSelection, the
selection data gets flagged with "newLineNeeded", if there is more than
one paragraph.

However, if not all of those paragraphs contribute to the actual
selection, this can lead to an unwanted leading newline.

Solution:
Remove the explicit SelectionData#markNewLine. There are no use-cases
where this would require this specific control over the newlines.
Instead, simply handle adding of newlines implicitly by checking the
number of already existing segments.


Change-Id: Iab7cfb2b8375da3da32440eb78084ef3a3348414
Signed-off-by: Sebastian Ratz <sebastian.ratz@sap.com>