Bug 538504 - Performance of FaceTextUtil.isShowingEntireContents()
For huge widgets avoid calculating all line heights to see whether
scrolling is necessary in the widget. Instead calculate the last visible
line and decide based on its index and visibility. This reduces the
costly line height calculation to a small number, completely decoupled
from the line count in the document.

In my profiling use case with a 42.000 lines git diff editor (which has
variable line height), the runtime of the changed method dropped from
640ms to less than 10ms, as measured in Yourkit.

Change-Id: I25cd71587afcf2c266ec48255cb87cb35da6cce2
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>