Bug 539154 - [IDE] Improve re-use of IDEApplication in extending classes
This change exposes a few methods of IDEApplication to implementers of
IDEApplication. Among these is the checking of workspace version against
IDE version.

This allows custom product IDE applications to also prompt for a
workspace location, as well as validate the chosen location in a manner
that is conform with the standard Eclipse IDE behavior. Furthermore its
possible to do more strict or less strict version checks.

Change-Id: I97bd77cd03cc1bd19f4f66f7bd83beeb4c7a2208
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>