Bug 535752 - Performance of search view
With many search results being available, the search view takes almost
all the time for refreshing the tree in the viewer. This can be avoided
since we know perfectly that there is no structural change, when calling
that method (the if insertChild() return value proves that). Therefore
instead of a structure refresh it is sufficient to update a single node
of the viewer only.

For a search with more than 500.000 results, this brings the search
runtime down from 3 minutes to 3 seconds on my system.

Change-Id: I9c80b9e2e0c6c9aaa8cc732c7eec2d76dfc55305
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>