Bug 539510: duplicate preference pages when using multiple windows
With the fix for bug 535679, quick access previous choices are partially
loaded in a job. If multiple windows exist, multiple such jobs can run
in parallel. This can result in undesired behavior, if loading a command
is not thread-safe.

This change ensures that the jobs loading previous choices do not run in
parallel, by adding a scheduling rule.

Change-Id: I5f9e657e0d5b414218177733eb274143206fc834
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>