Bug 407888 - BasicAliasTest.testDeepLink failed
The test creates a project with a symlink, tests some conditions, then
deletes the projects and asserts no symlink remains. Unfortunately,
AliasManager is notified of the delete via RefreshJob. So if the test
checks for aliases earlier than the refresh job, the test will fail.

This change adds a wait for the RefreshJob to run, after deleting the
test project. This avoids the race condition.

Change-Id: I3a721d97bf0cbf4cf324050750e8eb10d3c9d057
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>