Bug 536308: update AbstractContributionItem with delay
Use a temporal limit for the forced update instead of a counter.

The counter can be too low in situations where extremely many updates
calls are happening in a very short time.

It can also be too high in worst-case situations where for example a
single update call happens every 99 milliseconds.

Instead, remember the timestamp of the first update call that gets
queued and use the time difference to decide whether a forced update
needs to be performed.

Change-Id: I64fa3e262b2b44eca3574f0ebad1428d1ee72665
Signed-off-by: Sebastian Ratz <sebastian.ratz@sap.com>