Bug 541353 - Improvements for sporadically failing tests
This change adds extra waiting in the test CharsetTest.testBug207510.
The waiting is not time based but based on running jobs from
CharsetManager.

The change also adds the current stack trace whenever a missing delta is
detected by the test helper class ResourceDeltaVerifier.

And last, a lambda is converted to an inline class definition for easier
debugging.

Change-Id: I252e319ab048a4c7cbdb3c20df06b412687bd954
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>