Bug 23862 - [PropertiesView] Properties view doesn't update when switching between editors
Fix: Instead of showing properties for 'the last structural selection
that was made while the default properties view page was active',
properties will be showed for the current active part.

With this patch:
1. The default page is a message page, showing no properties available
   - Incorrect disposal of the old default page is fixed as with this
well
2. When a part provides properties, it always gets is own
PropertySheetPage. The page is either provided by the part or a default
properties page is created when the part is a ISelectionProvider.
3. The default message page is shown when a part is a
IContributedContentsView and there is not contributing part.
4. The PropertySheet does not listen to the global selection provider,
but only to the selection provider of the current active part.
5. When the selection to the PropertySheetPage is not a IStructured
selection or when the selection is empty, the part (or contributing
part) is provide as input to the viewer.
6. An adaptor is added to show the IFile properties when a IEditorPart
is provided as input to the properties viewer.

Change-Id: I877700815d6286d15619dc4b4e2fa6ab0041ba96
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>