Bug 540222 - All test projects should dump stack traces on timeout
When running tests with org.eclipse.test.EclipseTestRunner during
integration builds, stack traces and screenshots are provided if the
build times out (e.g. due to a hang in the tests). When running test
projects with maven however (e.g. when jdt.debug.tests run in a Gerrit
job), no such stack traces or screenshots are present. Debugging
sporadic hangs in Gerrit jobs is therefore very difficult.

This change is preparation for adding the dumps from EclipseTestRunner
to Gerrit jobs. For this we add the dump functionality at the location
of TestableObject, which is available in plug-ins which define test
applications.

The original code was moved from org.eclipse.test.EclipseTestRunner in
project org.eclipse.test of platform releng, with repository:
https://git.eclipse.org/r/platform/eclipse.platform.releng

The original code is mostly left untouched, aside from using logging
instead of system out and err streams for output.

Change-Id: I194e01de6ba4b7cbdedf4111a48d5f1345240d5d
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>