Bug 543746: Project Explorer should show explanatory text
Use a stack layout to switch between the "original" composite and an
additional composite given the user explanatory text. This text is
displayed when no projects are in the workspace. Once projects are
created a resource change listener switches back to the "original"
composite.

The explanatory text contains links to:
- project creation wizards according to the current perspective
- the "New Project Wizard" to allow creation of any type of project
If no perspective specific project creation wizards are found then a
simple text with a link to the "New Project Wizard" is shown.

A perspective listener takes care of re-freshing these links.

Change-Id: Ib7f3e7329f69e8ba540c877ac25d3aca4d69a862
Signed-off-by: Matthias Becker <ma.becker@sap.com>