Bug 543294 Only use StackParts belonging to the current area when creating a CTabFolder
When deciding to move the area's composite to a CTabFolder (due to
multiple stacks being created under the area) only consider Stacks that
are directly part of that area, or to SashContainers under it. But not
Stacks that are members of possibly CompositeParts inside that area.

Passed the fetched stack list to the rest of the methods that need to
check its minimisation or maximisation state.
Added AreaRendererTest.java with tests for the expected behaviour.

Change-Id: Ic82eb6563abf30d82c473d86a113ea5514e071c6
Signed-off-by: Ghaith Hachem<Ghaith.HACHEM@bachmann.info>