Bug 540817 - Suppress PropertyChangeEvent for identical working sets
On working set selection change two selection events are triggered - one
for the current selected menu item and one for the new one. The first
triggers a no-op change event, which calls
WorkingSetActionProvider#filterChangeListener. That listener resets the
viewer's frameList which leads to disposal of the menu. At the end the
second event is swallowed since the menu item was disposed already. By
suppressing the PropertyChangeEvent for no-op changes the too early
MenuItem disposal is avoided and the selection change event is
processed.

Change-Id: I19306ce5e3b97bb652358c2f916ca1ac9612165e
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>