Bug 543609 - Call saveState when a view part closes
Fixed the asymmetry in the workbench calling IViewPart.init(IViewSite,
IMemento) when a view part opens, but not calling saveState(IMemento)
when the view part closes within the running workbench session.

Modified WorkingSetTest to make testDeletedAndRecreated and
testMissingProjectsInWorkingSet pass again. They contained seemingly
needless calls to DisplayHelper.runEventLoop, which allowed the
asyncExec lambda scheduled in WorkingSetActionProvider.restoreState to
execute. This messed up the viewer input. The primary cause for the test
failures was the implicit assumption that saveState will not be called
on the tested navigator view. This assumption was wrong anyway because
of the workbench auto-save job, but calling saveState when the view is
closed in NavigatorTestBase.tearDown() made the tests fail predictably.

Change-Id: I2a5a898f4f1af326ba8cd02a0d4bd1a8e2c76d2a
Signed-off-by: Vladimir Piskarev <pisv@1c.ru>