Bug 543570 - Implement functionality to allow to match type children
Change markerHelp.exsd to contain the matchChildrenFlag, change
MarkerHelpRegistryReader to pass that on and change MarkerQuery to use
it.

Add a test for MarkerHelpRegistryReader, which checks, that the
matchChildren flag is read from the xml correctly.
Add a test for MarkerQuery, which checks, that the matchChildren flag is
used correctly and that with it enabled children are matched correctly.
Also add some extensions to the plugin.xml used by these tests.

Change-Id: Iadfd139ccc2b7924ec51e0d4b6ffd6b95a2dc621
Signed-off-by: Tim Neumann <Tim.Neumann@advantest.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>