Bug 541353 - Adjusted delta verification to handle multiple deltas
During CharsetTest, testBug207510 and testDeltaContainer are observed to
fail sporadically. One reason for this are periodic notifications sent
by the NotificationManager. Each workspace job will tell the
NotificationManager to schedule a notification, in particular
CharsetManager jobs will do so. It is possible for a notification from
the NotificationManager to run during the two tests and cause an
expected delta in a separate resource delta notification. Both test fail
in this case, as the verification code expects changes in exactly one
delta.

This change adjusts the two tests to either expect a single delta or
multiple deltas in their background delta verification. This is done by
building on top of the resource verification code, as to not change
behaviour in other unrelated tests.


Change-Id: Ibdfd8570875ab1123fb8c38268dbc4173a292dc0
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>