Bug 530450 - [Project Explorer] UI Freeze with nested projects
Change in marker used to trigger a refresh of the viewer, while an
element update should be invoked instead.
So we removed the code cascading markers changes to refresh from
NestedProjectContentProvider.
Instead, we made the NestedProjectsLabelProvider taking care of
listening to marker changes and reacting by updating its internal model.
Then, since the Label Provider doesn't have a reference to the viewer,
we made the contentprovider register itself for updates to the
labelprovider so the later invokes an updates of modified elements in
the viewer after its internal marker model updates.

Change-Id: Ib207a1bc8c7ff7a6aed85c2dc6e8ebe5420b1252
Signed-off-by: Mickael Istria <mistria@redhat.com>