Bug 378815 - Need better story for handling model elements that should disappear on startup
Use persisted state instead of transient data so that we allow model
elements to flag
themselves as non-persistent via their persistent data. This can also be
used for runtime elements.

This allows
model contributions via model fragments to mark certain parts as
"dynamic" in the sense that the model contribution is not present
anymore after uninstalling the contributing plug-in and a restart of the
application.


Change-Id: I1d09251490339a12195679fd140c8ddc833a7b3b
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>